+++ src/Files.App/Data/Items/FileTagsItemViewModel.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.UserControls.Widgets;
+using Files.App.ViewModels.UserControls.Widgets;
 using Files.Core.Storage;
 using Files.Core.Storage.Extensions;
