+++ src/Files.App/Data/Items/IWidgetCardItem.cs
@@ -2,9 +2,8 @@
 // Licensed under the MIT License. See the LICENSE.
 
 using Microsoft.UI.Xaml.Media.Imaging;
-using System.Threading.Tasks;
 
-namespace Files.App.UserControls.Widgets
+namespace Files.App.Data.Items
 {
 	public interface IWidgetCardItem<T>
 	{
