+++ src/Files.App/UserControls/Widgets/DrivesWidget.xaml
@@ -1,19 +1,15 @@
 <!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:HomePageWidget
+<UserControl
 	x:Class="Files.App.UserControls.Widgets.DrivesWidget"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:dataitems="using:Files.App.Data.Items"
 	xmlns:helpers="using:Files.App.Helpers"
-	xmlns:local="using:Files.App.UserControls.Widgets"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	mc:Ignorable="d">
 
-	<ItemsRepeater
-		x:Name="CardsList"
-		HorizontalAlignment="Stretch"
-		VerticalAlignment="Top"
-		ItemsSource="{x:Bind local:DrivesWidget.ItemsAdded, Mode=OneWay}">
+	<ItemsRepeater x:Name="CardsList" ItemsSource="{x:Bind ViewModel.ItemsAdded, Mode=OneWay}">
 
 		<ItemsRepeater.Layout>
 			<UniformGridLayout
@@ -26,7 +22,7 @@
 		</ItemsRepeater.Layout>
 
 		<ItemsRepeater.ItemTemplate>
-			<DataTemplate x:DataType="local:DriveCardItem">
+			<DataTemplate x:DataType="dataitems:DriveCardItem">
 				<Button
 					Padding="0"
 					HorizontalAlignment="Stretch"
@@ -56,6 +52,7 @@
 							<RowDefinition Height="*" />
 							<RowDefinition Height="*" />
 						</Grid.RowDefinitions>
+
 						<Image
 							Grid.RowSpan="3"
 							Grid.Column="0"
@@ -125,6 +122,7 @@
 							TextTrimming="CharacterEllipsis"
 							TextWrapping="NoWrap"
 							ToolTipService.ToolTip="{x:Bind Item.SpaceText, Mode=OneWay}" />
+
 					</Grid>
 				</Button>
 			</DataTemplate>
@@ -132,4 +130,4 @@
 
 	</ItemsRepeater>
 
-</local:HomePageWidget>
+</UserControl>
