+++ src/Files.App/UserControls/Widgets/FileTagsWidget.xaml
@@ -1,5 +1,5 @@
 ﻿<!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:HomePageWidget
+<UserControl
 	x:Class="Files.App.UserControls.Widgets.FileTagsWidget"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -7,11 +7,8 @@
 	xmlns:converters="using:Files.App.Converters"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:helpers="using:Files.App.Helpers"
-	xmlns:local="using:Files.App.UserControls.Widgets"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	xmlns:vm="using:Files.App.ViewModels.Widgets"
-	d:DesignHeight="300"
-	d:DesignWidth="400"
 	mc:Ignorable="d">
 
 	<UserControl.Resources>
@@ -19,7 +16,7 @@
 		<converters:ImageModelToImageConverter x:Key="ImageModelToImageConverter" />
 	</UserControl.Resources>
 
-	<ItemsRepeater ItemsSource="{x:Bind ViewModel.Containers}">
+	<ItemsRepeater ItemsSource="{x:Bind ViewModel.Containers, Mode=OneWay}">
 
 		<ItemsRepeater.Layout>
 			<UniformGridLayout
@@ -44,9 +41,7 @@
 					BorderThickness="1"
 					CornerRadius="{StaticResource ControlCornerRadius}">
 					<Grid.RowDefinitions>
-						<!--  Title  -->
 						<RowDefinition Height="Auto" />
-						<!--  Contents  -->
 						<RowDefinition Height="*" />
 					</Grid.RowDefinitions>
 
@@ -166,4 +161,4 @@
 
 	</ItemsRepeater>
 
-</local:HomePageWidget>
+</UserControl>
