+++ src/Files.App/UserControls/Widgets/QuickAccessWidget.xaml
@@ -1,17 +1,17 @@
 ﻿<!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:HomePageWidget
+<UserControl
 	x:Class="Files.App.UserControls.Widgets.QuickAccessWidget"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-	xmlns:local="using:Files.App.UserControls.Widgets"
+	xmlns:dataitems="using:Files.App.Data.Items"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	mc:Ignorable="d">
 
 	<ItemsRepeater
 		x:Name="CardsList"
 		HorizontalAlignment="Stretch"
-		ItemsSource="{x:Bind local:QuickAccessWidget.ItemsAdded, Mode=OneWay}">
+		ItemsSource="{x:Bind ViewModel.ItemsAdded, Mode=OneWay}">
 
 		<ItemsRepeater.Layout>
 			<UniformGridLayout
@@ -25,7 +25,7 @@
 		</ItemsRepeater.Layout>
 
 		<ItemsRepeater.ItemTemplate>
-			<DataTemplate x:DataType="local:FolderCardItem">
+			<DataTemplate x:DataType="dataitems:FolderCardItem">
 				<Button
 					Padding="8"
 					HorizontalAlignment="Stretch"
@@ -83,4 +83,4 @@
 
 	</ItemsRepeater>
 
-</local:HomePageWidget>
+</UserControl>
