+++ src/Files.App/UserControls/Widgets/RecentFilesWidget.xaml
@@ -1,15 +1,13 @@
 <!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:HomePageWidget
+<UserControl
 	x:Class="Files.App.UserControls.Widgets.RecentFilesWidget"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:helpers="using:Files.App.Helpers"
-	xmlns:local="using:Files.App.UserControls.Widgets"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
 	xmlns:recent="using:Files.App.Utils.RecentItem"
-	Background="{ThemeResource SolidBackgroundFillColorBaseBrush}"
 	mc:Ignorable="d">
 
 	<StackPanel x:Name="RecentsListGrid" Orientation="Vertical">
@@ -17,7 +15,7 @@
 		<!--  Message when recent files are disabled in File Explorer  -->
 		<muxc:InfoBar
 			x:Name="RecentFilesDisabledOnWindowsWarning"
-			x:Load="{x:Bind IsRecentFilesDisabledInWindows, Mode=OneWay}"
+			x:Load="{x:Bind ViewModel.IsRecentFilesDisabledInWindows, Mode=OneWay}"
 			IsClosable="False"
 			IsIconVisible="True"
 			IsOpen="True"
@@ -28,7 +26,7 @@
 		<TextBlock
 			x:Name="RecentItemDescription"
 			HorizontalAlignment="Stretch"
-			x:Load="{x:Bind IsEmptyRecentsTextVisible, Mode=OneWay}"
+			x:Load="{x:Bind ViewModel.IsEmptyRecentsTextVisible, Mode=OneWay}"
 			FontSize="14"
 			Text="{helpers:ResourceString Name=RecentItemDescription/Text}"
 			TextAlignment="Center"
@@ -43,7 +41,7 @@
 			IsRightTapEnabled="True"
 			ItemClick="RecentsView_ItemClick"
 			ItemContainerTransitions="{x:Null}"
-			ItemsSource="{x:Bind recentItemsCollection}"
+			ItemsSource="{x:Bind ViewModel.Items, Mode=OneWay}"
 			RightTapped="ListView_RightTapped"
 			SelectionMode="None">
 
@@ -122,4 +120,4 @@
 
 	</StackPanel>
 
-</local:HomePageWidget>
+</UserControl>
