+++ src/Files.App/ViewModels/UserControls/Widgets/WidgetCardItem.cs
@@ -1,12 +1,12 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.UserControls.Widgets
+namespace Files.App.ViewModels.UserControls.Widgets
 {
 	public abstract class WidgetCardItem : ObservableObject
 	{
-		public virtual string Path { get; set; }
+		public virtual string? Path { get; set; }
 
-		public virtual object Item { get; set; }
+		public virtual object? Item { get; set; }
 	}
 }
